requirements.txt
lightrag-hku
python-dotenv
ollama
google-genai
langchain
langchain-text-splitters
langchain-community
jq
unstructured
markdown
esprima
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
nltk
numpy>=1.22.4,<1.29.0
